requirements.txt
nibabel==5.2.*
numpy==1.23.*
pynrrd==1.0.*
pytest==8.3.*
pytest-cov==5.0.*
pytest-mock==3.14.*
pytest-metadata==3.1.*
pytest-console-scripts==1.4.*
pytest-html==4.1.*
pyyaml==6.0.*